scrap/requirements.txt
asgiref @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_edg6nzjrx0/croots/recipe/asgiref_1661287549316/work
boto3==1.26.83
botocore==1.29.83
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
decorator==5.1.1
Django @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_066uop7vqa/croots/recipe/django_1661292540858/work
django-filter==22.1
django-phonenumber-field==7.0.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
dynaconf==3.1.11
hvac==1.0.2
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.2
phonenumbers==8.13.6
Pillow==9.4.0
pycparser==2.21
pyhcl==0.4.4
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sqlparse @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_204sy4e5t2/croot/sqlparse_1668793448973/work
urllib3==1.26.14
validators==0.20.0
